move events related code to ln_events
